requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
certifi==2017.4.17
channels==1.1.5
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
dj-static==0.0.6
Django==1.11.2
django-ckeditor==5.3.0
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-easy-select2==1.3.4
django-js-asset==0.1.1
django-material==1.0.0
django-rest-auth==0.9.1
django-smart-selects==1.5.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
gunicorn==19.7.1
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
msgpack-python==0.4.8
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pystache==0.5.4
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
static==1.1.1
static3==0.7.0
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
websocket-client==0.44.0
zope.interface==4.4.2